% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_config}
\alias{add_config}
\alias{add_user}
\alias{add_group}
\alias{add_key}
\title{Configure a Zotero API call}
\usage{
add_config(.data, ..., user_id, group_id, api_key)

add_user(.data, user_id)

add_group(.data, group_id)

add_key(.data, api_key)
}
\arguments{
\item{.data}{An object of class \code{zotero}}

\item{...}{fields to be pulled from \code{zotero_config} (optional)}

\item{user_id}{Optionally specify the user ID}

\item{group_id}{Optionally specify the group ID}

\item{api_key}{Optionally specify the API key}
}
\description{
Functions to allow the user to specify which identifiers should be passed to
an API call.
}
\details{
\code{add_config()} is a catch-all function that tries to do sensible things
when passing user data to a \code{zotero} call. Alternatively, the microfunctions
\code{add_user()}, \code{add_group()} and \code{add_key()} each add individual pieces of
information, with fewer checks.

Regardless of which function is used, specifying any of \code{user_id}, \code{group_id}
or \code{api_key} will overrule use of the defaults as stored in \code{zotero_config()}.
In such cases, however, the defaults stored in \code{zotero_config} will not
themselves be overwritten.
}
