% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Collect data from Zotero}
\usage{
collect(x, path)
}
\arguments{
\item{x}{An object of class `zotero`}

\item{path}{An (optional) path to download files to}
}
\description{
Function to catch data from pipes, pass to Zotero API, and return a tibble
}
