% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.zotero}
\alias{collect.zotero}
\title{Collect data from Zotero}
\usage{
\method{collect}{zotero}(.data, path)
}
\arguments{
\item{.data}{An object of class \code{zotero}}

\item{path}{An (optional) path to download files to}
}
\description{
Function to catch data from pipes, pass to Zotero API, and return a tibble
}
