% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.zotero}
\alias{filter.zotero}
\title{Filter a Zotero query}
\usage{
\method{filter}{zotero}(.data, ...)
}
\arguments{
\item{.data}{An object of class `zotero`}

\item{...}{additional
importFrom glue glue}
}
\description{
Test port from {dplyr}
}
