% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Add a group to a Zotero API call}
\usage{
group(x, id)
}
\arguments{
\item{x}{An object of class `zotero`}

\item{id}{User ID. Can be left blank if zotero_config(group_id) has been set.}
}
\description{
Designed to be piped following a `zotero()` call.
}
