% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups}
\alias{groups}
\title{Add a group to a Zotero API call}
\usage{
groups(.data)
}
\arguments{
\item{.data}{An object of class \code{zotero}}
}
\description{
Designed to be piped following a \code{zotero()} call.
}
