% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{items}
\alias{items}
\title{Add item data to a Zotero API call}
\usage{
items(.data, id)
}
\arguments{
\item{.data}{An object of class \code{zotero}}

\item{id}{An (optional) id to look up}
}
\description{
First attempt
}
