% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tags}
\alias{tags}
\title{Add tag data to a Zotero API call}
\usage{
tags(.data, id)
}
\arguments{
\item{.data}{An object of class \code{zotero}}

\item{id}{An (optional) id to look up}
}
\description{
First attempt
}
