% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{Add a user to a Zotero API call}
\usage{
user(x, id)
}
\arguments{
\item{x}{An object of class `zotero`}

\item{id}{User ID. Can be left blank if zotero_config(user_id) has been set.}
}
\description{
Designed to be piped following a `zotero()` call.
}
