% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/zotero.R
\name{print.zotero}
\alias{print.zotero}
\alias{zotero}
\title{Start a pipe for querying Zotero}
\usage{
\method{print}{zotero}(.data, ...)

zotero()
}
\arguments{
\item{.data}{An object of class `zotero`}

\item{...}{Any further arguments (currently ignored)}
}
\description{
This function creates an object of class 'zotero', which can then be used
to pipe arguments to construct and enact a query
}
