% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_config.R
\name{zotero_config}
\alias{zotero_config}
\title{Configure a zotero session}
\usage{
zotero_config(user_id = NULL, group_id = NULL, api_key = NULL)
}
\arguments{
\item{x}{a list of configurable options}
}
\description{
Simple config stuff
}
