% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boobook-package.R
\docType{package}
\name{boobook}
\alias{boobook}
\title{Tools for calling the Zotero API (v3)}
\description{
Zotero (https://www.zotero.org) is a bibliographic management tool. It
includes an online version whereby data can be accessed through an API.

To import data exported from a desktop version of Zotero - where files are
stored locally - try using \code{{synthesisr}} instead.
}
\section{Functions}{

\itemize{

\item{\code{\link[=zotero_config]{zotero_config()}} Store information on user or group IDs, or API keys}
\item{\code{\link[=zotero]{zotero()}} Start a pipe}
\item{\code{\link[=configure]{configure()}} Specify which identifiers to pass to the API}
\item{\code{\link[=collect.zotero]{collect.zotero()}} Make an API call}
\item{\code{\link[=groups]{groups()}}}
\item{\code{\link[=items]{items()}}}
\item{\code{\link[=collections]{collections()}}}
\item{\code{\link[=tags]{tags()}}}
\item{\code{\link[=filter.zotero]{filter.zotero()}} Filtering (may not work properly yet)}
}
}

