% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Add item data to a Zotero API call}
\usage{
collections(.data, id)
}
\arguments{
\item{.data}{An object of class \code{zotero}}

\item{id}{An (optional) id to look up}
}
\description{
First attempt
}
