% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/configure.R
\name{add_user}
\alias{add_user}
\alias{add_group}
\alias{add_key}
\alias{configure}
\title{Configure a Zotero API call}
\usage{
add_user(.data, user_id)

add_group(.data, group_id)

add_key(.data, api_key)

configure(.data, ..., user_id, group_id, api_key)
}
\arguments{
\item{.data}{An object of class `zotero`}

\item{user_id}{Optionally specify the user ID}

\item{group_id}{Optionally specify the group ID}

\item{api_key}{Optionally specify the API key}

\item{...}{fields to be pulled from `zotero_config` (optional)}
}
\description{
Functions to allow the user to specify which identifiers should be passed to
an API call.
}
\details{
`configure()` is a catch-all function that tries to do sensible things
when passing user data to a `zotero` call. Alternatively, the microfunctions
`add_user()`, `add_group()` and `add_key()` each add individual pieces of
information, with fewer checks.

Regardless of which function is used, specifying any of `user_id`, `group_id`
or `api_key` will overrule use of the defaults as stored in `zotero_config()`.
In such cases, however, the defaults stored in `zotero_config` will not
themselves be overwritten.
}
