% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_config.R
\name{zotero_config}
\alias{zotero_config}
\title{Configure a zotero session}
\usage{
zotero_config(user_id = NULL, group_id = NULL, api_key = NULL, clear = FALSE)
}
\arguments{
\item{user_id}{Unique identifier for users}

\item{group_id}{Unique identifier for groups}

\item{api_key}{A Zotero API key}

\item{clear}{(logical) Should the cache be cleared? This will forget all previous arguments}
}
\description{
Simple config stuff
}
\details{
This function automatically converts supplied arguments to strings (class `character`).
}
